requirements.txt
streamlit
pillow
numpy
scikit-learn
opencv-python
matplotlib
plotly
streamlit-drawable-canvas